requirements.txt
pika==1.1.0
pika-stubs==0.1.3
Pillow==7.2.0
